\name{jordan}
\docType{data}
\alias{jordan}
\title{Data from Jordan subset of PISA 2006 data}
\description{
  The data stem from the large-scale assessment study PISA 2006
  (Organisation for Economic Co-Operation and Development, 2009) where
  competencies of 15-year-old students in reading, mathematics, and science
  are assessed using nationally representative samples in 3-year cycles. In
  this example, data from the student background questionnaire from the
  Jordan sample of PISA 2006 were used. Only data of students with complete
  responses to all 15 items (N = 6,038) were considered.
}
\usage{data(jordan)}
\format{
  A data frame of fifteen variables and 6,038 observations:
  \describe{
    \item{x1}{indicator for academic self-concept in science.}
    \item{x2}{indicator for academic self-concept in science.}
    \item{x3}{indicator for academic self-concept in science.}
    \item{x4}{indicator for academic self-concept in science.}
    \item{x5}{indicator for academic self-concept in science.}
    \item{x6}{indicator for enjoyment of science.}
    \item{x7}{indicator for enjoyment of science.}
    \item{x8}{indicator for enjoyment of science.}
    \item{x9}{indicator for enjoyment of science.}
    \item{x10}{indicator for enjoyment of science.}
    \item{x11}{indicator for enjoyment of science.}
    \item{y1}{indicator for career aspirations in science.}
    \item{y2}{indicator for career aspirations in science.}
    \item{y3}{indicator for career aspirations in science.}
    \item{y4}{indicator for career aspirations in science.}
  }
}
\source{
Organisation for Economic Co-Operation and Development (2009). \emph{Pisa 2006:
Science competencies for tomorrow's world} (Tech. Rep.). Paris, France.
Obtained from: \url{http://pisa2006.acer.edu.au/downloads.php}
}
\examples{
data(jordan)
}
\keyword{datasets}

